Setup the base for an upgrade/ability system
Changed the functionality of gears to function as an exp bar, which when full is supposed to provide upgrades and abilities. Added some UI for the gear/exp bar.

Made a basic stat system which uses scriptable objects and runtime stats.
This serves as a base for the upcoming upgrade system.

Also got rid of the old upgrade system/shop from the main menu. That will be later converted into a meta progression shop for permanent upgrades.